chore: moving omnichannel e2e tests to its own folder
