+++ apps/meteor/tests/e2e/omnichannel/omnichannel-canned-responses-sidebar.spec.ts
@@ -1,12 +1,15 @@
 import { faker } from '@faker-js/faker';
 import type { Page } from '@playwright/test';
 
-import { createAuxContext } from './fixtures/createAuxContext';
-import { Users } from './fixtures/userStates';
-import { OmnichannelLiveChat, HomeChannel } from './page-objects';
-import { test } from './utils/test';
+import { IS_EE } from '../config/constants';
+import { createAuxContext } from '../fixtures/createAuxContext';
+import { Users } from '../fixtures/userStates';
+import { OmnichannelLiveChat, HomeChannel } from '../page-objects';
+import { test } from '../utils/test';
 
-test.describe('Omnichannel contact info', () => {
+test.describe('Omnichannel Canned Responses Sidebar', () => {
+	test.skip(!IS_EE, 'Enterprise Only');
+
 	let poLiveChat: OmnichannelLiveChat;
 	let newUser: { email: string; name: string };
 
@@ -35,7 +38,7 @@ test.describe('Omnichannel contact info', () => {
 		await agent.page.close();
 	});
 
-	test('Receiving a message from visitor, and seeing its information', async ({ page }) => {
+	test('Receiving a message from visitor', async ({ page }) => {
 		await test.step('Expect send a message as a visitor', async () => {
 			await page.goto('/livechat');
 			await poLiveChat.btnOpenLiveChat('R').click();
@@ -48,9 +51,9 @@ test.describe('Omnichannel contact info', () => {
 			await agent.poHomeChannel.sidenav.getSidebarItemByName(newUser.name).click();
 		});
 
-		await test.step('Expect to be see contact information and edit', async () => {
-			await agent.poHomeChannel.content.btnContactInformation.click();
-			await agent.poHomeChannel.content.btnContactEdit.click();
+		await test.step('Expect to be able to open canned responses sidebar and creation', async () => {
+			await agent.poHomeChannel.content.btnCannedResponses.click();
+			await agent.poHomeChannel.content.btnNewCannedResponse.click();
 		});
 	});
 });
