+++ apps/meteor/tests/e2e/omnichannel/omnichannel-custom-fields.spec.ts
@@ -1,6 +1,6 @@
-import { Users } from './fixtures/userStates';
-import { OmnichannelCustomFields } from './page-objects';
-import { test, expect } from './utils/test';
+import { Users } from '../fixtures/userStates';
+import { OmnichannelCustomFields } from '../page-objects';
+import { test, expect } from '../utils/test';
 
 test.use({ storageState: Users.admin.state });
