+++ apps/meteor/tests/e2e/omnichannel/omnichannel-priorities.spec.ts
@@ -1,9 +1,9 @@
 import { faker } from '@faker-js/faker';
 
-import { IS_EE } from './config/constants';
-import { Users } from './fixtures/userStates';
-import { OmnichannelPriorities } from './page-objects/omnichannel-priorities';
-import { test, expect } from './utils/test';
+import { IS_EE } from '../config/constants';
+import { Users } from '../fixtures/userStates';
+import { OmnichannelPriorities } from '../page-objects/omnichannel-priorities';
+import { test, expect } from '../utils/test';
 
 const PRIORITY_NAME = faker.person.firstName();
