+++ apps/meteor/tests/e2e/omnichannel/omnichannel-send-transcript.spec.ts
@@ -1,20 +1,17 @@
 import { faker } from '@faker-js/faker';
 import type { Page } from '@playwright/test';
 
-import { IS_EE } from './config/constants';
-import { createAuxContext } from './fixtures/createAuxContext';
-import { Users } from './fixtures/userStates';
-import { OmnichannelLiveChat, HomeChannel } from './page-objects';
-import { test } from './utils/test';
+import { IS_EE } from '../config/constants';
+import { createAuxContext } from '../fixtures/createAuxContext';
+import { Users } from '../fixtures/userStates';
+import { OmnichannelLiveChat, HomeChannel } from '../page-objects';
+import { test, expect } from '../utils/test';
 
-test.describe('Omnichannel Canned Responses Sidebar', () => {
-	test.skip(!IS_EE, 'Enterprise Only');
-
+test.describe('omnichannel-transcript', () => {
 	let poLiveChat: OmnichannelLiveChat;
 	let newUser: { email: string; name: string };
 
 	let agent: { page: Page; poHomeChannel: HomeChannel };
-
 	test.beforeAll(async ({ api, browser }) => {
 		newUser = {
 			name: faker.person.firstName(),
@@ -51,9 +48,18 @@ test.describe('Omnichannel Canned Responses Sidebar', () => {
 			await agent.poHomeChannel.sidenav.getSidebarItemByName(newUser.name).click();
 		});
 
-		await test.step('Expect to be able to open canned responses sidebar and creation', async () => {
-			await agent.poHomeChannel.content.btnCannedResponses.click();
-			await agent.poHomeChannel.content.btnNewCannedResponse.click();
+		await test.step('Expect to be able to send transcript to email', async () => {
+			await agent.poHomeChannel.content.btnSendTranscript.click();
+			await agent.poHomeChannel.content.btnSendTranscriptToEmail.click();
+			await agent.poHomeChannel.content.btnModalConfirm.click();
+			await expect(agent.poHomeChannel.toastSuccess).toBeVisible();
+		});
+
+		await test.step('Expect to be not able send transcript as PDF', async () => {
+			test.skip(!IS_EE, 'Enterprise Only');
+			await agent.poHomeChannel.content.btnSendTranscript.click();
+			await agent.poHomeChannel.content.btnSendTranscriptAsPDF.hover();
+			await expect(agent.poHomeChannel.content.btnSendTranscriptAsPDF).toHaveAttribute('aria-disabled', 'true');
 		});
 	});
 });
